Closes #115
